+++ src/Blaster.WebApi/CreatedAtRouteResultConverter.cs
@@ -1,0 +1,25 @@
+﻿using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.Infrastructure;
+
+namespace Blaster.WebApi
+{
+    public class CreatedAtRouteResultConverter<T> : IConvertToActionResult
+    {
+        private readonly string _routeName;
+        private readonly object _routeValues;
+
+        public CreatedAtRouteResultConverter(string routeName, object routeValues, T value)
+        {
+            _routeName = routeName;
+            _routeValues = routeValues;
+            Value = value;
+        }
+
+        public T Value { get; }
+
+        public IActionResult Convert()
+        {
+            return new CreatedAtRouteResult(_routeName, _routeValues, Value);
+        }
+    }
+}
